Bulk deletion integration UI to API
