+++ GUI/src/pages/Agency/Agency.tsx
@@ -75,12 +75,6 @@ const Agency: FC = () => {
   const [addUrlModal, setAddUrlModal] = useState(false);
   const [editModal, setEditModal] = useState<Source | null>(null);
   const [deleteModal, setDeleteModal] = useState<Source | null>(null);
-  
-  // Bulk action confirmation dialogs
-  const [bulkDeleteConfirm, setBulkDeleteConfirm] = useState<Row<Source>[] | null>(null);
-  const [bulkRefreshConfirm, setBulkRefreshConfirm] = useState<Row<Source>[] | null>(null);
-  const [bulkIncludeConfirm, setBulkIncludeConfirm] = useState<Row<Source>[] | null>(null);
-  const [bulkExcludeConfirm, setBulkExcludeConfirm] = useState<Row<Source>[] | null>(null);
 
   // Add table state for server-side pagination and sorting
   const [pagination, setPagination] = useState<PaginationState>({
@@ -460,55 +454,6 @@ const Agency: FC = () => {
     refreshMutation.mutate(sourceId);
   };
 
-  // Multiselect action handlers
-  const handleBulkDelete = (selectedRows: Row<Source>[]) => {
-    setBulkDeleteConfirm(selectedRows);
-  };
-
-  const confirmBulkDelete = () => {
-    if (!bulkDeleteConfirm) return;
-    const selectedIds = bulkDeleteConfirm.map(row => row.original.baseId);
-    console.log('Delete selected rows:', selectedIds);
-    // TODO: Implement bulk delete API call
-    setBulkDeleteConfirm(null);
-  };
-
-  const handleBulkRefresh = (selectedRows: Row<Source>[]) => {
-    setBulkRefreshConfirm(selectedRows);
-  };
-
-  const confirmBulkRefresh = () => {
-    if (!bulkRefreshConfirm) return;
-    const selectedIds = bulkRefreshConfirm.map(row => row.original.baseId);
-    console.log('Refresh selected rows:', selectedIds);
-    // TODO: Implement bulk refresh API call
-    setBulkRefreshConfirm(null);
-  };
-
-  const handleBulkInclude = (selectedRows: Row<Source>[]) => {
-    setBulkIncludeConfirm(selectedRows);
-  };
-
-  const confirmBulkInclude = () => {
-    if (!bulkIncludeConfirm) return;
-    const selectedIds = bulkIncludeConfirm.map(row => row.original.baseId);
-    console.log('Include selected rows:', selectedIds);
-    // TODO: Implement bulk include API call
-    setBulkIncludeConfirm(null);
-  };
-
-  const handleBulkExclude = (selectedRows: Row<Source>[]) => {
-    setBulkExcludeConfirm(selectedRows);
-  };
-
-  const confirmBulkExclude = () => {
-    if (!bulkExcludeConfirm) return;
-    const selectedIds = bulkExcludeConfirm.map(row => row.original.baseId);
-    console.log('Exclude selected rows:', selectedIds);
-    // TODO: Implement bulk exclude API call
-    setBulkExcludeConfirm(null);
-  };
-
   // Handle pagination change
   const handlePaginationChange = (newPagination: PaginationState) => {
     setPagination(newPagination);
@@ -520,38 +465,6 @@ const Agency: FC = () => {
   };
 
   const columns: ColumnDef<Source>[] = [
-    {
-      id: 'select',
-      header: ({ table }) => {
-        const checkboxRef = useRef<HTMLInputElement>(null);
-        
-        useEffect(() => {
-          if (checkboxRef.current) {
-            checkboxRef.current.indeterminate = table.getIsSomeRowsSelected() && !table.getIsAllRowsSelected();
-          }
-        }, [table.getIsSomeRowsSelected(), table.getIsAllRowsSelected()]);
-
-        return (
-          <input
-            ref={checkboxRef}
-            type="checkbox"
-            checked={table.getIsAllRowsSelected()}
-            onChange={table.getToggleAllRowsSelectedHandler()}
-          />
-        );
-      },
-      cell: ({ row }) => (
-        <input
-          type="checkbox"
-          checked={row.getIsSelected()}
-          disabled={!row.getCanSelect()}
-          onChange={row.getToggleSelectedHandler()}
-        />
-      ),
-      meta: {
-        size: 10,
-      },
-    },
     {
       accessorKey: 'url',
       header: t('knowledgeBase.url'),
@@ -689,34 +602,6 @@ const Agency: FC = () => {
     },
   ];
 
-  // Define multiselect actions
-  const multiselectActions: MultiselectAction[] = [
-    {
-      label: t('global.include'),
-      icon: <MdCheckCircle />,
-      variant: 'secondary',
-      onClick: handleBulkInclude,
-    },
-    {
-      label: t('global.exclude'),
-      icon: <MdCancel />,
-      variant: 'secondary',
-      onClick: handleBulkExclude,
-    },
-    {
-      label: t('knowledgeBase.refresh'),
-      icon: <MdRefresh />,
-      variant: 'secondary',
-      onClick: handleBulkRefresh,
-    },
-    {
-      label: t('global.delete'),
-      icon: <MdOutlineDeleteOutline />,
-      variant: 'danger',
-      onClick: handleBulkDelete,
-    }
-  ];
-
   // Show loading state
   if (isLoadingAgency || isLoadingSources) {
     return <div>Loading...</div>;
@@ -763,8 +648,6 @@ const Agency: FC = () => {
           filterable
           pagesCount={sourcesData?.totalPages ?? 0}
           isClientSide={false}
-          enableRowSelection={true}
-          multiselectActions={multiselectActions}
         />
 
         <div className="agencies__footer">
@@ -951,121 +834,6 @@ const Agency: FC = () => {
         </Dialog>
       )}
 
-      {/* Bulk Delete Confirmation Modal */}
-      {bulkDeleteConfirm && (
-        <Dialog
-          title={t('global.delete')}
-          onClose={() => setBulkDeleteConfirm(null)}
-          footer={
-            <Track gap={16} justify="end">
-              <Button
-                appearance="secondary"
-                onClick={() => setBulkDeleteConfirm(null)}
-              >
-                {t('global.cancel')}
-              </Button>
-              <Button
-                appearance="error"
-                onClick={confirmBulkDelete}
-              >
-                {t('global.delete')}
-              </Button>
-            </Track>
-          }
-        >
-          {t('global.confirmBulkDelete', {
-            count: bulkDeleteConfirm.length,
-            unit: bulkDeleteConfirm.length === 1 ? t('global.source') : t('global.sources')
-          })}
-        </Dialog>
-      )}
-
-      {/* Bulk Refresh Confirmation Modal */}
-      {bulkRefreshConfirm && (
-        <Dialog
-          title={t('knowledgeBase.refresh')}
-          onClose={() => setBulkRefreshConfirm(null)}
-          footer={
-            <Track gap={16} justify="end">
-              <Button
-                appearance="secondary"
-                onClick={() => setBulkRefreshConfirm(null)}
-              >
-                {t('global.cancel')}
-              </Button>
-              <Button
-                appearance="primary"
-                onClick={confirmBulkRefresh}
-              >
-                {t('knowledgeBase.refresh')}
-              </Button>
-            </Track>
-          }
-        >
-          {t('global.confirmBulkRefresh', {
-            count: bulkRefreshConfirm.length,
-            unit: bulkRefreshConfirm.length === 1 ? t('global.source') : t('global.sources')
-          })}
-        </Dialog>
-      )}
-
-      {/* Bulk Include Confirmation Modal */}
-      {bulkIncludeConfirm && (
-        <Dialog
-          title={t('global.includeSources')}
-          onClose={() => setBulkIncludeConfirm(null)}
-          footer={
-            <Track gap={16} justify="end">
-              <Button
-                appearance="secondary"
-                onClick={() => setBulkIncludeConfirm(null)}
-              >
-                {t('global.cancel')}
-              </Button>
-              <Button
-                appearance="primary"
-                onClick={confirmBulkInclude}
-              >
-                {t('global.include')}
-              </Button>
-            </Track>
-          }
-        >
-          {t('global.confirmBulkInclude', {
-            count: bulkIncludeConfirm.length,
-            unit: bulkIncludeConfirm.length === 1 ? t('global.source') : t('global.sources')
-          })}
-        </Dialog>
-      )}
-
-      {/* Bulk Exclude Confirmation Modal */}
-      {bulkExcludeConfirm && (
-        <Dialog
-          title={t('global.excludeSources')}
-          onClose={() => setBulkExcludeConfirm(null)}
-          footer={
-            <Track gap={16} justify="end">
-              <Button
-                appearance="secondary"
-                onClick={() => setBulkExcludeConfirm(null)}
-              >
-                {t('global.cancel')}
-              </Button>
-              <Button
-                appearance="primary"
-                onClick={confirmBulkExclude}
-              >
-                {t('global.exclude')}
-              </Button>
-            </Track>
-          }
-        >
-          {t('global.confirmBulkExclude', {
-            count: bulkExcludeConfirm.length,
-            unit: bulkExcludeConfirm.length === 1 ? t('global.source') : t('global.sources')
-          })}
-        </Dialog>
-      )}
     </div>
   );
 };
