+++ GUI/translations/en/common.json
@@ -39,10 +39,10 @@
     "excludeSources": "Exclude Sources",
     "source": "source",
     "sources": "sources",
-    "confirmBulkDelete": "Are you sure you want to delete {count} selected {unit}?",
-    "confirmBulkRefresh": "Are you sure you want to refresh {count} selected {unit}?",
-    "confirmBulkInclude": "Are you sure you want to include {count} selected {unit}?",
-    "confirmBulkExclude": "Are you sure you want to exclude {count} selected {unit}?",
+    "confirmBulkDelete": "Are you sure you want to delete {{count}} selected {{unit}}?",
+    "confirmBulkRefresh": "Are you sure you want to refresh {{count}} selected {{unit}}?",
+    "confirmBulkInclude": "Are you sure you want to include {{count}} selected {{unit}}?",
+    "confirmBulkExclude": "Are you sure you want to exclude {{count}} selected {{unit}}?",
     "name": "Name",
     "idCode": "ID code",
     "status": "Status",
