+++ GUI/translations/et/common.json
@@ -39,10 +39,10 @@
     "excludeSources": "Välista allikad",
     "source": "allikas",
     "sources": "allikat",
-    "confirmBulkDelete": "Kas oled kindel, et soovid kustutada {count} valitud {unit}?",
-    "confirmBulkRefresh": "Kas oled kindel, et soovid värskendada {count} valitud {unit}?",
-    "confirmBulkInclude": "Kas oled kindel, et soovid kaasata {count} valitud {unit}?",
-    "confirmBulkExclude": "Kas oled kindel, et soovid välistada {count} valitud {unit}?",
+    "confirmBulkDelete": "Kas oled kindel, et soovid kustutada {{count}} valitud {{unit}}?",
+    "confirmBulkRefresh": "Kas oled kindel, et soovid värskendada {{count}} valitud {{unit}}?",
+    "confirmBulkInclude": "Kas oled kindel, et soovid kaasata {{count}} valitud {{unit}}?",
+    "confirmBulkExclude": "Kas oled kindel, et soovid välistada {{count}} valitud {{unit}}?",
     "name": "Nimi",
     "idCode": "Isikukood",
     "status": "Staatus",
